Introduce a CMake-based build system
This change removes prepare-depends.sh and Makefile, replacing it with
a CMake-based build system and git submodules of the needed
dependencies.

Notably, this simplifies the build of libsnark to be just
  git clone --recursive git://github.com/scipr-lab/libsnark.git
  cd libsnark
  mkdir build && cd build && cmake .. && make

Since CMake also supports Ninja (apt-get install ninja-build on Debian
and derivatives), an even faster build could be obtained by using
  cmake -GNinja .. && ninja

README.md has been updated to reflect the new build system.

All options supported by the Makefile are also supported by the CMake.

As an added benefit, users of libsnark can embed libsnark as a git
submodule and add it to their CMakeLists.txt, without having to
install it.

This change is a rebase of the earlier patch sent to the scipr-lab
developers, updated for new features introduced since then.