Ensure that all files compile & sort includes.
I have added a couple missing includes, so that every *.cpp or *.hpp
file in the repository can now be compiled individually. I also grouped
all includes (by stdlib/external libs/libff) and sorted them.

instruction_packing_gadget was deleted because it did not compile, was
clearly incompatible with the current codebase, and was not used
anywhere. According to Madars, it was used in a previous version of
the reduction.